+++ NEWS.md
@@ -1,3 +1,8 @@
+# 0.13.2
+
+3D drc graph entropy calculations implemented in c++.
+Network function will now only use decently expressed TFs (<10% of cells) and apply a lower correlation cutoff (0.1 new vs 0.8 old)
+
 # 0.10.1
 
 export2cellexalvr: The hull datasets are no longer created